refactor: rename request_info.py to client_ip.py
